Reorganized file structure - each component now has it's own folder for jsx and css Fix - sorting mode now resets when swapping to favorites or watched
